requirements.txt
socket
threading
python-dotenv
